pyproject.toml
# DEPENDENCIES

[tool.poetry]
name = "regex-finder"
version = "0.0.0"
description = "Algorithm that tries to find regex based on given strings that match this unknown regex."
authors = ["Piotr Karaś <pkaras@student.agh.edu.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
automata-lib = "^8.3.0"
coloraide = "^3.3.1"
numpy = "^1.26.4"
poetry-dynamic-versioning = "^1.3.0"
pygraphviz = "^1.13"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
fawltydeps = "^0.13.3"
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.3"
isort = {version = "^5.12.0", extras = ["colors"]}
mypy = "^1.3.0"
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"


# CONFIGURATION

[tool.black]
line-length = 120

[tool.coverage.run]
branch = true
omit = [
    "*test*",
    "*__init__*",
]

[tool.coverage.report]
exclude_lines = [
    "if TYPE_CHECKING:",
    "raise NotImplementedError()",
]

[tool.fawltydeps]
code = ["src"]
deps = ["pyproject.toml"]
ignore_unused = [
    # Needed for visualisation in automata-lib
    "coloraide",
    "pygraphviz",
    # Dev dependencies. This list should contain ALL of them!
    "black",
    "fawltydeps",
    "flake8",
    "flake8-pyproject",
    "isort",
    "mypy",
    "poetry-dynamic-versioning",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.flake8]
# S101 == Use of assert detected. The enclosed code will be removed when compiling to optimised byte code.
# W503 == line break before binary operator. Conflicts with W504 - line break after binary operator.
# A003 == . Conflicts with W504 - line break after binary operator.
# E203 == whitespace before ':' Conflicts with black
ignore = ['S101', 'W503', 'A003', 'E203']
max-line-length = 120

[tool.isort]
profile = "black"
# isort configuration:
# https://github.com/timothycrosley/isort/wiki/isort-Settings
# ignored bc: https://github.com/PyCQA/isort/issues/1640 Workaround run isort with params isort --line-length=80
line_length = 120
order_by_type = false

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
show_traceback = true
no_implicit_optional = true
check_untyped_defs = true
strict_equality = true

[tool.poetry-dynamic-versioning]
enable = true
bump = true

# pytest settings
[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--strict-markers --cov=. --cov-fail-under=90.0 --cov-config=.coveragerc"
pythonpath = [
    ".",
]
asyncio_mode = "auto"
